Refactor ruleset
