+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wazuh",
-  "version": "3.11.0",
+  "version": "3.10.2",
   "revision": "0561",
   "code": "0561-0",
   "kibana": {
