+++ public/controllers/management/components/management/ruleset/decoder-info.js
@@ -17,6 +17,7 @@ import {
 import { connect } from 'react-redux';
 
 import RulesetHandler from './utils/ruleset-handler';
+import { colors } from './utils/colors';
 
 
 import {
@@ -25,32 +26,24 @@ import {
   cleanInfo,
   updateFilters,
   cleanFilters
-} from '../../../../redux/actions/rulesetActions';
+} from '../../../../../redux/actions/rulesetActions';
 
-class WzRuleInfo extends Component {
+class WzDecoderInfo extends Component {
   constructor(props) {
     super(props);
-    this.complianceEquivalences = {
-      pci: 'PCI DSS',
-      gdpr: 'GDPR',
-      gpg13: 'GPG 13',
-      hipaa: 'HIPAA',
-      'nist-800-53': 'NIST-800-53'
-    }
 
     this.rulesetHandler = RulesetHandler;
     this.columns = [
       {
-        field: 'id',
-        name: 'ID',
+        field: 'name',
+        name: 'Name',
         align: 'left',
         sortable: true,
-        width: '5%',
         render: value => {
           return (
-            <EuiToolTip position="top" content={`Show rule ID ${value} information`}>
+            <EuiToolTip position="top" content={`Show ${value} decoder information`}>
               <EuiLink onClick={() => {
-                this.changeBetweenRules(value);
+                this.changeBetweenDecoders(value);
               }
               }>
                 {value}
@@ -60,98 +53,51 @@ class WzRuleInfo extends Component {
         }
       },
       {
-        field: 'description',
-        name: 'Description',
+        field: 'details.program_name',
+        name: 'Program name',
         align: 'left',
-        sortable: true,
-        width: '30%'
+        sortable: true
       },
       {
-        field: 'groups',
-        name: 'Groups',
+        field: 'details.order',
+        name: 'Order',
         align: 'left',
-        sortable: true,
-        width: '10%'
-      },
-      {
-        field: 'pci',
-        name: 'PCI',
-        align: 'left',
-        sortable: true,
-        width: '10%'
-      },
-      {
-        field: 'gdpr',
-        name: 'GDPR',
-        align: 'left',
-        sortable: true,
-        width: '10%'
-      },
-      {
-        field: 'hipaa',
-        name: 'HIPAA',
-        align: 'left',
-        sortable: true,
-        width: '10%'
-      },
-      {
-        field: 'nist-800-53',
-        name: 'NIST 800-53',
-        align: 'left',
-        sortable: true,
-        width: '10%'
-      },
-      {
-        field: 'level',
-        name: 'Level',
-        align: 'left',
-        sortable: true,
-        width: '5%'
+        sortable: true
       },
       {
         field: 'file',
         name: 'File',
         align: 'left',
         sortable: true,
-        width: '15%',
         render: (value, item) => {
           return (
             <EuiToolTip position="top" content={`Show ${value} content`}>
               <EuiLink onClick={async () => {
                 const noLocal = item.path.startsWith('ruleset/');
-                const result = await this.rulesetHandler.getRuleContent(value, noLocal);
+                const result = await this.rulesetHandler.getDecoderContent(value, noLocal);
                 const file = { name: value, content: result, path: item.path };
                 this.props.updateFileContent(file);
               }
-              }>
-                {value}
-              </EuiLink>
+              }>{value}</EuiLink>
             </EuiToolTip>
           )
         }
+      },
+      {
+        field: 'path',
+        name: 'Path',
+        align: 'left',
+        sortable: true
       }
     ];
   }
+
 
   componentWillUnmount() {
     // When the component is going to be unmounted its info is clear
     this.props.cleanInfo();
   }
-
-  /**
-   * Build an object with the compliance info about a rule
-   * @param {Object} ruleInfo 
-   */
-  buildCompliance(ruleInfo) {
-    const compliance = {};
-    const complianceKeys = ['gdpr', 'gpg13', 'hipaa', 'nist-800-53', 'pci'];
-    Object.keys(ruleInfo).forEach(key => {
-      if (complianceKeys.includes(key) && ruleInfo[key].length) compliance[key] = ruleInfo[key]
-    });
-    return compliance || {};
-  }
-
-
+  
   /**
    * Clean the existing filters and sets the new ones and back to the previous section
    */
@@ -164,24 +110,14 @@ class WzRuleInfo extends Component {
 
   /**
    * Render the basic information in a list
-   * @param {Number} id 
-   * @param {Number} level 
+   * @param {Number} position 
    * @param {String} file 
    * @param {String} path 
    */
-  renderInfo(id, level, file, path) {
+  renderInfo(position, file, path) {
     return (
       <ul>
-        <li key="id"><b>ID:</b>&nbsp;{id}</li>
-        <EuiSpacer size="s" />
-        <li key="level"><b>Level:</b>
-          <EuiToolTip position="top" content={`Filter by this level: ${level}`}>
-            <EuiLink onClick={async () => this.setNewFiltersAndBack({ level: level })}>
-              &nbsp;{level}
-            </EuiLink>
-          </EuiToolTip>
-        </li>
-
+        <li key="position"><b>Position:</b>&nbsp;<span className="subdued-color">{position}</span></li>
         <EuiSpacer size="s" />
         <li key="file"><b>File:</b>
           <EuiToolTip position="top" content={`Filter by this file: ${file}`}>
@@ -198,7 +134,6 @@ class WzRuleInfo extends Component {
             </EuiLink>
           </EuiToolTip>
         </li>
-
         <EuiSpacer size="s" />
       </ul>
     )
@@ -206,14 +141,22 @@ class WzRuleInfo extends Component {
 
   /**
    * Render a list with the details
-* @param {Array} details
-    */
+   * @param {Array} details 
+   */
   renderDetails(details) {
     const detailsToRender = [];
     Object.keys(details).forEach(key => {
+      let content = details[key]
+      if (key === 'regex') {
+        content = this.colorRegex(content);
+      } else if (key === 'order') {
+        content = this.colorOrder(content);
+      } else {
+        content = <span className="subdued-color">{details[key]}</span>;
+      }
       detailsToRender.push(
         <Fragment>
-          <li key={key}><b>{key}:</b>&nbsp;{details[key]}</li>
+          <li key={key}><b>{key}:</b>&nbsp;{content}</li>
           <EuiSpacer size="s" />
         </Fragment>
       );
@@ -226,85 +169,51 @@ class WzRuleInfo extends Component {
   }
 
   /**
-   * Render the groups
-* @param {Array} groups
-    */
-  renderGroups(groups) {
-    const listGroups = [];
-    groups.forEach(group => {
-      listGroups.push(
-        <Fragment>
-          <EuiLink onClick={async () => this.setNewFiltersAndBack({ group: group })}>
-            <EuiToolTip position="top" content={`Filter by this group: ${group}`}>
-              <li key={group}>
-                {group}
-              </li>
-            </EuiToolTip>
-          </EuiLink>
-          <EuiSpacer size="s" />
-        </Fragment>
-      );
-    });
-    return (
-      <ul>
-        {listGroups}
-      </ul>
-    )
-  }
-
-  /**
-   * Render the compliance(HIPAA, NIST...)
-* @param {Array} compliance
-    */
-  renderCompliance(compliance) {
-    const listCompliance = [];
-    const keys = Object.keys(compliance);
-    for (let i in Object.keys(keys)) {
-      const key = keys[i];
-      listCompliance.push(
-        <Fragment>
-          <li key={key}><b>{this.complianceEquivalences[key]}</b></li>
-          <EuiSpacer size="s" />
-        </Fragment>
-      )
-      compliance[key].forEach(element => {
-        const filters = {};
-        filters[key] = element;
-        listCompliance.push(
-          <Fragment>
-            <EuiLink onClick={async () => this.setNewFiltersAndBack({ filters })}>
-              <EuiToolTip position="top" content="Filter by this compliance">
-                <li key={element}>{element}</li>
-              </EuiToolTip>
-            </EuiLink>
-            <EuiSpacer size="s" />
-          </Fragment>
-        );
-      });
+   * This set a color to a given order
+   * @param {String} order
+   */
+  colorOrder(order) {
+    order = order.toString();
+    let valuesArray = order.split(',');
+    const result = [];
+    for (let i = 0, len = valuesArray.length; i < len; i++) {
+      const coloredString = <span style={{ color: colors[i] }}>{valuesArray[i]}</span>;
+      result.push(coloredString);
     }
-    return (
-      <ul>
-        {listCompliance}
-      </ul>
-    )
+    return result;
   }
 
   /**
-   * Changes between rules
-* @param {Number} ruleId
-    */
-  changeBetweenRules(ruleId) {
-    this.setState({ currentRuleId: ruleId });
+     * This set a color to a given regex
+     * @param {String} regex
+     */
+  colorRegex(regex) {
+    regex = regex.toString();
+    const starts = <span className="subdued-color">{regex.split('(')[0]}</span>;
+    let valuesArray = regex.match(/\(((?!<\/span>).)*?\)(?!<\/span>)/gim);
+    const result = [starts];
+    for (let i = 0, len = valuesArray.length; i < len; i++) {
+      const coloredString = <span style={{ color: colors[i] }}>{valuesArray[i]}</span>;
+      result.push(coloredString);
+    }
+    return result;
+  }
+
+  /**
+   * Changes between decoders
+   * @param {Number} name 
+   */
+  changeBetweenDecoders(name) {
+    this.setState({ currentDecoder: name });
   }
 
   render() {
-    const { ruleInfo, isLoading } = this.props.state;
-    const currentRuleId = (this.state && this.state.currentRuleId) ? this.state.currentRuleId : ruleInfo.current;
-    const rules = ruleInfo.items;
-    const currentRuleArr = rules.filter(r => { return r.id === currentRuleId });
-    const currentRuleInfo = currentRuleArr[0];
-    const { description, details, file, path, level, id, groups } = currentRuleInfo;
-    const compliance = this.buildCompliance(currentRuleInfo);
+    const { decoderInfo, isLoading } = this.props.state;
+    const currentDecoder = (this.state && this.state.currentDecoder) ? this.state.currentDecoder : decoderInfo.current;
+    const decoders = decoderInfo.items;
+    const currentDecoderArr = decoders.filter(r => { return r.name === currentDecoder });
+    const currentDecoderInfo = currentDecoderArr[0];
+    const { position, details, file, name, path } = currentDecoderInfo;
     const columns = this.columns;
 
 
@@ -312,12 +221,12 @@ class WzRuleInfo extends Component {
       <EuiPage style={{ background: 'transparent' }}>
         <EuiFlexGroup>
           <EuiFlexItem>
-            {/* Rule description name */}
+            {/* Decoder description name */}
             <EuiFlexGroup>
               <EuiFlexItem grow={false}>
                 <EuiTitle>
                   <h2>
-                    <EuiToolTip position="right" content="Back to rules">
+                    <EuiToolTip position="right" content="Back to decoders">
                       <EuiButtonIcon
                         aria-label="Back"
                         color="subdued"
@@ -325,7 +234,7 @@ class WzRuleInfo extends Component {
                         iconType="arrowLeft"
                         onClick={() => this.props.cleanInfo()} />
                     </EuiToolTip>
-                    {description}
+                    {name}
                   </h2>
                 </EuiTitle>
               </EuiFlexItem>
@@ -339,7 +248,7 @@ class WzRuleInfo extends Component {
                   <EuiText color="subdued">Information</EuiText>
                   <EuiSpacer size="xs" className="subdued-background" />
                   <EuiSpacer size="s" />
-                  {this.renderInfo(id, level, file, path)}
+                  {this.renderInfo(position, file, path)}
                 </EuiPanel>
               </EuiFlexItem>
               {/* Details */}
@@ -351,26 +260,6 @@ class WzRuleInfo extends Component {
                   {this.renderDetails(details)}
                 </EuiPanel>
               </EuiFlexItem>
-              {/* Groups */}
-              <EuiFlexItem>
-                <EuiPanel paddingSize="s">
-                  <EuiText color="subdued">Groups</EuiText>
-                  <EuiSpacer size="xs" className="subdued-background" />
-                  <EuiSpacer size="s" />
-                  {this.renderGroups(groups)}
-                </EuiPanel>
-              </EuiFlexItem>
-              {/* Compliance */}
-              {Object.keys(compliance).length > 0 && (
-                <EuiFlexItem>
-                  <EuiPanel paddingSize="s">
-                    <EuiText color="subdued">Compliance</EuiText>
-                    <EuiSpacer size="xs" className="subdued-background" />
-                    <EuiSpacer size="s" />
-                    {this.renderCompliance(compliance)}
-                  </EuiPanel>
-                </EuiFlexItem>
-              )}
             </EuiFlexGroup>
 
             {/* Table */}
@@ -382,7 +271,7 @@ class WzRuleInfo extends Component {
                     <EuiFlexItem>
                       <EuiTitle size="s">
                         <h5>
-                          Related rules
+                          Related decoders
                       </h5>
                       </EuiTitle>
                     </EuiFlexItem>
@@ -392,12 +281,12 @@ class WzRuleInfo extends Component {
                     <EuiFlexItem>
                       <EuiInMemoryTable
                         itemId="id"
-                        items={rules}
+                        items={decoders}
                         columns={columns}
                         pagination={true}
                         loading={isLoading}
                         sorting={true}
-                        message={false}
+                        message={null}
                       />
                     </EuiFlexItem>
                   </EuiFlexGroup>
@@ -428,4 +317,4 @@ const mapDispatchToProps = (dispatch) => {
 };
 
 
-export default connect(mapStateToProps, mapDispatchToProps)(WzRuleInfo);
+export default connect(mapStateToProps, mapDispatchToProps)(WzDecoderInfo);
