+++ public/controllers/management/components/management/ruleset/section-selector.js
