+++ public/controllers/management/components/management/ruleset/utils/colors.js
