+++ public/controllers/management/components/management/ruleset/utils/columns.js
