+++ public/controllers/management/components/management/ruleset/utils/valid-configuration.js
@@ -10,7 +10,7 @@
  * Find more information about this on the LICENSE file.
  */
 
-import { WzRequest } from '../../../../../react-services/wz-request';
+import { WzRequest } from '../../../../../../react-services/wz-request';
 
 const validateConfigAfterSent = async (node = false) => {
   try {
