+++ public/less/common.less
@@ -16,6 +16,10 @@
 
 /* Custom healthcheck and blank screen styles */
 
+.kbnGlobalBannerList{
+  display: none;
+}
+
 .error-notify {
   font-size: 20px;
   color: black;
