+++ public/redux/actions/managementActions.js
@@ -1,0 +1,7 @@
+
+export const updateManagementSection = (section) => {
+  return {
+    type: 'UPDATE_MANAGEMENT_SECTION',
+    section
+  }
+}
