+++ public/redux/reducers/rootReducers.js
@@ -12,9 +12,10 @@
 
 import { combineReducers } from 'redux';
 import rulesetReducers from './rulesetReducers';
-
+import managementReducers from './managementReducers';
 
 export default combineReducers({
-  rulesetReducers
+  rulesetReducers,
+  managementReducers
 })
