+++ public/templates/management/groups/groups.html
@@ -11,7 +11,7 @@
 
         <!-- Groups table -->
         <div ng-show="!gp.load" layout="column" ng-if="(!gp.currentGroup || !gp.currentGroup.name) && gp.groupsTableProps && !gp.lookingGroup">
-            <react-component name="GroupsTable" props="gp.groupsTableProps" />
+<!--             <react-component name="GroupsTable" props="gp.groupsTableProps" /> -->
         </div>
         
         <!-- Inside the group -->
