+++ public/templates/management/management.html
@@ -1,0 +1,3 @@
+<div ng-if="!rctrl.loading && mctrl.managementProps" layout="column">  
+  <react-component name="WzManagement" props="managementProps"></react-component>
+</div>
