+++ public/templates/management/management.pug
@@ -5,6 +5,10 @@ include ./configuration/configuration.pug
 include ./monitoring/monitoring.pug
 include ./logs.html
 include ./reporting.html
+
+include ./management.html
+
 include ./groups/groups.html
 include ./ruleset/ruleset.pug
+
 include ../footer.foot
