+++ public/templates/management/ruleset/decoders/decoders.pug
@@ -1,4 +1,3 @@
-include ./decoders.head
 include ./decoders-list.html
 include ./decoders-detail.html
 include ../../../footer.foot
