--- public/templates/management/ruleset/rules/rules-list.html
@@ -1,3 +1,0 @@
-<div ng-if="!rctrl.loading && !rctrl.viewingDetail" layout="column">  
-  <react-component name="WzRuleset" props="rctrl.mainRulesProps"></react-component>
-</div>
