+++ public/templates/management/ruleset/rules/rules.pug
@@ -1,4 +1,3 @@
 include ./rules.head
-include ./rules-list.html
 include ./rules-detail.html
 include ../../../footer.foot
