--- public/templates/management/ruleset/ruleset.foot
@@ -1,1 +1,0 @@
-</div>
