+++ public/templates/management/ruleset/ruleset.pug
@@ -1,8 +1,3 @@
-include ./ruleset.head
-include ./rules/rules.pug
-include ./decoders/decoders.pug
-include ./cdblists/cdblists.pug
-include ./files/files.pug
+include ./ruleset.html
 include ./logtest.html
-include ./ruleset.foot
 include ../../footer.foot
